ENH: Refactor F2PY to follow a logical skeleton
