+++ numpy/f2py/__main__.py
@@ -1,5 +1,6 @@
 # See:
 # https://web.archive.org/web/20140822061353/http://cens.ioc.ee/projects/f2py2e
-from numpy.f2py.f2py2e import main
+import sys
+from numpy.f2py.frontend.f2py2e import main
 
-main()
+sys.exit(main())
