+++ numpy/f2py/codegen/cfuncs.py
@@ -17,7 +17,7 @@ Pearu Peterson
 import sys
 import copy
 
-from . import __version__
+from numpy.f2py import __version__
 
 f2py_version = __version__.version
 errmess = sys.stderr.write
@@ -1351,7 +1351,7 @@ capi_fail:
 
 
 def buildcfuncs():
-    from .capi_maps import c2capi_map
+    from numpy.f2py.stds.pyf.capi_maps import c2capi_map
     for k in c2capi_map.keys():
         m = 'pyarr_from_p_%s1' % k
         cppmacros[
