+++ numpy/f2py/frontend/crackfortran.py
@@ -147,13 +147,13 @@ import os
 import copy
 import platform
 
-from . import __version__
+from numpy.f2py import __version__
 
 # The environment provided by auxfuncs.py is needed for some calls to eval.
 # As the needed functions cannot be determined by static inspection of the
 # code, it is safest to use import * pending a major refactoring of f2py.
-from .auxfuncs import *
-from . import symbolic
+from numpy.f2py.stds.auxfuncs import *
+from numpy.f2py.stds import symbolic
 
 f2py_version = __version__.version
