--- numpy/f2py/setup.cfg
@@ -1,3 +1,0 @@
-[bdist_rpm]
-doc_files = docs/
-            tests/
