+++ numpy/f2py/stds/auxfuncs.py
@@ -19,8 +19,8 @@ import sys
 import types
 from functools import reduce
 
-from . import __version__
-from . import cfuncs
+from numpy.f2py import __version__
+from numpy.f2py.codegen import cfuncs
 
 __all__ = [
     'applyrules', 'debugcapi', 'dictappend', 'errmess', 'gentitle',
@@ -648,7 +648,7 @@ def getcallprotoargument(rout, cb_map={}):
         outmess(
             'warning: callstatement is defined without callprotoargument\n')
         return
-    from .capi_maps import getctype
+    from numpy.f2py.stds.pyf.capi_maps import getctype
     arg_types, arg_types2 = [], []
     if l_and(isstringfunction, l_not(isfunction_wrap))(rout):
         arg_types.extend(['char*', 'size_t'])
