+++ numpy/f2py/stds/f77/common_rules.py
@@ -13,15 +13,15 @@ $Date: 2005/05/06 10:57:33 $
 Pearu Peterson
 
 """
-from . import __version__
+from numpy.f2py import __version__
 f2py_version = __version__.version
 
-from .auxfuncs import (
+from numpy.f2py.stds.auxfuncs import (
     hasbody, hascommon, hasnote, isintent_hide, outmess
 )
-from . import capi_maps
-from . import func2subr
-from .crackfortran import rmbadname
+from numpy.f2py.stds.pyf import capi_maps
+from numpy.f2py.codegen import func2subr
+from numpy.f2py.frontend.crackfortran import rmbadname
 
 
 def findcommonblocks(block, top=1):
