+++ numpy/f2py/stds/f90/f90mod_rules.py
@@ -19,14 +19,14 @@ f2py_version = 'See `f2py -v`'
 
 import numpy as np
 
-from . import capi_maps
-from . import func2subr
-from .crackfortran import undo_rmbadname, undo_rmbadname1
+from numpy.f2py.stds.pyf import capi_maps
+from numpy.f2py.codegen import func2subr
+from numpy.f2py.frontend.crackfortran import undo_rmbadname, undo_rmbadname1
 
 # The environment provided by auxfuncs.py is needed for some calls to eval.
 # As the needed functions cannot be determined by static inspection of the
 # code, it is safest to use import * pending a major refactoring of f2py.
-from .auxfuncs import *
+from numpy.f2py.stds.auxfuncs import *
 
 options = {}
 
@@ -85,7 +85,7 @@ fgetdims2_sa = """\
 
 
 def buildhooks(pymod):
-    from . import rules
+    from numpy.f2py.codegen import rules
     ret = {'f90modhooks': [], 'initf90modhooks': [], 'body': [],
            'need': ['F_FUNC', 'arrayobject.h'],
            'separatorsfor': {'includes0': '\n', 'includes': '\n'},
