+++ numpy/f2py/stds/pyf/cb_rules.py
@@ -13,8 +13,8 @@ $Date: 2005/07/20 11:27:58 $
 Pearu Peterson
 
 """
-from . import __version__
-from .auxfuncs import (
+from numpy.f2py import __version__
+from numpy.f2py.stds.auxfuncs import (
     applyrules, debugcapi, dictappend, errmess, getargs, hasnote, isarray,
     iscomplex, iscomplexarray, iscomplexfunction, isfunction, isintent_c,
     isintent_hide, isintent_in, isintent_inout, isintent_nothide,
@@ -22,7 +22,7 @@ from .auxfuncs import (
     isstringfunction, issubroutine, l_and, l_not, l_or, outmess, replace,
     stripcomma, throw_error
 )
-from . import cfuncs
+from numpy.f2py.codegen import cfuncs
 
 f2py_version = __version__.version
