+++ numpy/f2py/stds/symbolic.py
