+++ numpy/f2py/tests/test_crackfortran.py
@@ -1,9 +1,12 @@
-import pytest
-import numpy as np
-from numpy.f2py.crackfortran import markinnerspaces
-from . import util
-from numpy.f2py import crackfortran
 import textwrap
+
+import pytest
+
+import numpy as np
+from numpy.f2py.frontend.crackfortran import markinnerspaces
+from numpy.f2py.frontend import crackfortran
+
+from . import util
 
 
 class TestNoSpace(util.F2PyTest):
