+++ numpy/f2py/tests/test_kind.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 
-from numpy.f2py.crackfortran import (
+from numpy.f2py.frontend.crackfortran import (
     _selected_int_kind_func as selected_int_kind,
     _selected_real_kind_func as selected_real_kind,
 )
