+++ numpy/f2py/tests/test_regression.py
@@ -1,6 +1,7 @@
 import os
 import pytest
 
+import numpy.f2py
 import numpy as np
 
 from . import util
@@ -60,7 +61,7 @@ class TestNumpyVersionAttribute(util.F2PyTest):
 
 
 def test_include_path():
-    incdir = np.f2py.get_include()
+    incdir = numpy.f2py.get_include()
     fnames_in_dir = os.listdir(incdir)
     for fname in ("fortranobject.c", "fortranobject.h"):
         assert fname in fnames_in_dir
