+++ numpy/f2py/utils/diagnose.py
